+++ src/app/app.module.ts
@@ -49,15 +49,15 @@ import { ReqeustDirectorComponent } from './components/pages/requests/views/dire
 import { ReqeustWildaComponent } from './components/pages/requests/views/wilda-panel/reqeust-wilda/reqeust-wilda.component';
 import { RequestRectorComponent } from './components/pages/requests/views/rector-panel/request-rector/request-rector.component';
 import { RejectDialogComponent } from './components/shared/reject-dialog/reject-dialog.component';
-import {SecurityService} from './components/shared/security/SecurityService';
-import {SafeHttpClient} from './components/shared/security/SafeHttpClient';
-import {LoginService} from './components/shared/security/LoginService';
-import { UserService } from './components/shared/security/roles/UserService';
-import { WildaService } from './components/shared/security/roles/WildaService';
-import { DirectorService } from './components/shared/security/roles/DirectorService';
-import { RectorService } from './components/shared/security/roles/RectorService';
-import { LocalStorageService } from './components/shared/security/LocalStorageService';
-import {HttpErrorInterceptor} from './components/shared/security/HttpErrorInterceptor';
+import {SecurityService} from './services/security/SecurityService';
+import {SafeHttpClient} from './services/security/SafeHttpClient';
+import {LoginService} from './services/security/LoginService';
+import { UserService } from './services/security/roles/UserService';
+import { WildaService } from './services/security/roles/WildaService';
+import { DirectorService } from './services/security/roles/DirectorService';
+import { RectorService } from './services/security/roles/RectorService';
+import { LocalStorageService } from './services/security/LocalStorageService';
+import {HttpErrorInterceptor} from './services/security/HttpErrorInterceptor';
 import {PaginatorIntl} from './extra/paginator-intl/paginator-intl';
 import {NgxSpinnerModule} from 'ngx-spinner';
 import { YesNoDialogComponent } from './components/shared/yes-no-dialog/yes-no-dialog.component';
