+++ src/app/components/pages/personal-data/personal-data.component.ts
@@ -8,8 +8,8 @@ import {find} from 'rxjs/operators';
 import {RequestDataService} from '../../../services/request-data.service';
 import {DialogService} from '../../../services/dialog.service';
 import {RestService} from '../../../services/rest-service';
-import {SecurityService} from '../../shared/security/SecurityService';
-import {LocalStorageService} from '../../shared/security/LocalStorageService';
+import {SecurityService} from '../../../services/security/SecurityService';
+import {LocalStorageService} from '../../../services/security/LocalStorageService';
 import {NgxSpinnerService} from 'ngx-spinner';
 import {Result} from '../../shared/yes-no-dialog/yes-no-dialog.component';
 import {PersonalDataService} from '../../../services/personal-data.service';
