+++ src/app/services/security/roles/DirectorService.ts
@@ -1,6 +1,6 @@
 import {ActivatedRouteSnapshot, CanActivate, RouterStateSnapshot} from '@angular/router';
 import {Injectable} from '@angular/core';
-import {PersonalDataService} from '../../../../services/personal-data.service';
+import {PersonalDataService} from '../../personal-data.service';
 import {Observable} from 'rxjs';
